move distributions class out from policy_util
