+++ test/lib/test_distribution.py
@@ -1,5 +1,5 @@
 from flaky import flaky
-from slm_lab.agent.algorithm import policy_util
+from slm_lab.lib import distribution
 import pytest
 import torch
 
@@ -10,7 +10,7 @@ import torch
 def test_argmax(pdparam_type):
     pdparam = torch.tensor([1.1, 10.0, 2.1])
     # test both probs or logits
-    pd = policy_util.Argmax(**{pdparam_type: pdparam})
+    pd = distribution.Argmax(**{pdparam_type: pdparam})
     for _ in range(10):
         assert pd.sample().item() == 1
     assert torch.equal(pd.probs, torch.tensor([0., 1., 0.]))
@@ -22,7 +22,7 @@ def test_argmax(pdparam_type):
 ])
 def test_gumbel_categorical(pdparam_type):
     pdparam = torch.tensor([1.1, 10.0, 2.1])
-    pd = policy_util.GumbelCategorical(**{pdparam_type: pdparam})
+    pd = distribution.GumbelCategorical(**{pdparam_type: pdparam})
     for _ in range(10):
         assert torch.is_tensor(pd.sample())
 
@@ -36,7 +36,7 @@ def test_multicategorical(pdparam_type):
     pdparam2 = torch.tensor([0.0, 0.0, 10.0])
     pdparams = [pdparam0, pdparam1, pdparam2]
     # use a probs
-    pd = policy_util.MultiCategorical(**{pdparam_type: pdparams})
+    pd = distribution.MultiCategorical(**{pdparam_type: pdparams})
     assert isinstance(pd.probs, list)
     # test probs only since if init from logits, probs will be close but not precise
     if pdparam_type == 'probs':
